+++ internal/tasks/util/doc.go
@@ -1,3 +1,3 @@
-// Package taskutil contains type definitions and helper functions for working with Tasks & Tools
+// Package taskutil contains type definitions and functionality for working with Tasks & Tools
 // across the oscar codebase.
 package taskutil
