+++ internal/tasks/util/system.go
@@ -82,62 +82,6 @@ func InitSystem(ctx context.Context) error {
 	return nil
 }
 
-// NewRun returns a populated [Run].
-func NewRun(ctx context.Context, runType string) (Run, error) {
-	// Handle system init
-	if err := InitSystem(ctx); err != nil {
-		return Run{}, fmt.Errorf("initializing system: %w", err)
-	}
-
-	return Run{
-		Type:      runType,
-		StartTime: time.Now(),
-		Failures:  make([]string, 0),
-	}, nil
-}
-
-// PrintRunTypeBanner prints a banner about the type of [Run] underway.
-func (run Run) PrintRunTypeBanner() {
-	padding := 9
-	fmt.Printf("%s\n", strings.Repeat("@", len(run.Type)+padding))
-	fmt.Printf("@ Run: %s @\n", run.Type)
-	fmt.Printf("%s\n\n", strings.Repeat("@", len(run.Type)+padding))
-}
-
-// PrintTaskMapBanner prints a banner about the [TaskMap] being run.
-func (run Run) PrintTaskMapBanner(lang string) {
-	fmt.Printf(
-		"=== %s %s>\n",
-		lang, strings.Repeat("=", 64-len(lang)),
-	)
-}
-
-// PrintTaskBanner prints a banner about the Task being run.
-func (run Run) PrintTaskBanner(task Tasker) {
-	// NOTE: no trailing newline on purpose
-	fmt.Printf("> %s %s............", task.InfoText(), strings.Repeat(".", 32-len(task.InfoText())))
-}
-
-// ReportSuccess prints information about the success of a [Run].
-func (run Run) ReportSuccess() {
-	fmt.Printf("All tasks succeeded! (%s)\n\n", RunDurationString(run.StartTime))
-}
-
-// ReportFailure prints information about the failure of a [Run]. It takes an `error` arg in case
-// the caller is expecting to return a joined error because of e.g. deferred calls or later-checked
-// errors that an outer variable already holds.
-func (run Run) ReportFailure(err error) error {
-	iprint.Errorf("\n%s\n", strings.Repeat("=", 65))
-	iprint.Errorf("The following tasks failed: (%s)\n", RunDurationString(run.StartTime))
-	for _, f := range run.Failures {
-		iprint.Errorf("- %s\n", f)
-	}
-	iprint.Errorf("%s\n\n", strings.Repeat("=", 65))
-
-	err = errors.Join(err, errors.New("one or more tasks failed"))
-	return err
-}
-
 // RunCommand takes a string slice containing a command & its args to run, and returns a consistent
 // error message in case of failure. It also returns the command output, in case the caller needs to
 // parse it on their own.
@@ -166,67 +110,16 @@ func RunCommand(ctx context.Context, cmdArgs []string) (string, error) {
 	return strings.TrimSuffix(string(output), "\n"), nil
 }
 
-// GetRepoComposition returns a populated [Repo].
-func GetRepoComposition(ctx context.Context) (Repo, error) {
-	var errs error
-
-	hasGo, err := filesExistInTree(ctx, GetFileTypeListerCommand("go"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasPython, err := filesExistInTree(ctx, GetFileTypeListerCommand("py"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasShell, err := filesExistInTree(ctx, GetFileTypeListerCommand("sh"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasTerraform, err := filesExistInTree(ctx, GetFileTypeListerCommand("tf"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasContainerfile, err := filesExistInTree(ctx, GetFileTypeListerCommand("containerfile"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasYaml, err := filesExistInTree(ctx, GetFileTypeListerCommand("yaml"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	hasMarkdown, err := filesExistInTree(ctx, GetFileTypeListerCommand("md"))
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-
-	if errs != nil {
-		return Repo{}, errs
-	}
-
-	repo := Repo{
-		HasGo:            hasGo,
-		HasPython:        hasPython,
-		HasShell:         hasShell,
-		HasTerraform:     hasTerraform,
-		HasContainerfile: hasContainerfile,
-		HasYaml:          hasYaml,
-		HasMarkdown:      hasMarkdown,
-	}
-	iprint.Debugf("repo composition: %+v\n", repo)
-
-	return repo, nil
+// RunDurationString returns a calculated duration used to indicate how long a particular Task (or
+// set of Tasks) took to run.
+func RunDurationString(t time.Time) string {
+	return fmt.Sprintf("t: %s", time.Since(t).Round(time.Second/1000).String())
 }
 
-// GetFileTypeListerCommand takes a [ripgrep]-known file type, and returns a slice containing a
-// command & its args to find matching files. ripgrep is used as the default file-finder across the
-// codebase because not only is it fast, but it also supports files like `.gitignore` without any
-// extra configuration.
+// GetFileTypeListerCommand takes a [ripgrep]-known file type, and returns a string (to be used as
+// an arg after `bash -c`) representing a command & its args to find matching files. ripgrep is used
+// as the default file-finder across the codebase because not only is it fast, but it also supports
+// files like `.gitignore` without any extra configuration.
 //
 // [ripgrep]: https://github.com/BurntSushi/ripgrep
 func GetFileTypeListerCommand(fileType string) string {
@@ -239,31 +132,6 @@ func GetFileTypeListerCommand(fileType string) string {
 	default:
 		return fmt.Sprintf(`rg --files --type '%s' || true`, fileType)
 	}
-}
-
-// RunDurationString returns a calculated duration used to indicate how long a particular Task (or
-// set of Tasks) took to run.
-func RunDurationString(t time.Time) string {
-	return fmt.Sprintf("t: %s", time.Since(t).Round(time.Second/1000).String())
-}
-
-// RenderRunCommandArgs uses [Tool.RunArgs] and does naive templating to replace certain values
-// before being used.
-//
-// This is useful because when instantiating [Tasker]s, sometimes the fields need to be
-// self-referential within the struct. For example, if a [Tool.RunArgs] needs to specify a config
-// file path, but it can't know that value until instantiation (even though it's likely defined
-// right below it in [Tool.ConfigFilePath]), you can write the `RunArgs` to have a
-// `{{ConfigFilePath}}` placeholder that will be interpolated when calling this function.
-func (t Tool) RenderRunCommandArgs() []string {
-	out := make([]string, len(t.RunArgs))
-	for i, arg := range t.RunArgs {
-		out[i] = strings.ReplaceAll(arg, `{{ConfigFilePath}}`, t.ConfigFilePath)
-	}
-
-	iprint.Debugf("RenderRunCommandArgs: %#v\n", out)
-
-	return out
 }
 
 // installMise installs [mise] into [consts.OscarHomeBin], if not found there.
